BioInfoWeb/requirements.txt
# Main
flask
sqlalchemy
flask-socketio
pymysql
werkzeug

# In api
ViennaRNA
biopython